stm32: renamed pwr_common_all to pwr_common_v1, and pwr_common_l01 to pwr_common_v2
