+++ include/libopencm3/stm32/common/pwr_common_v1.h
@@ -30,8 +30,8 @@ specific memorymap.h header before including this header file.*/
 /** @cond */
 #ifdef LIBOPENCM3_PWR_H
 /** @endcond */
-#ifndef LIBOPENCM3_PWR_COMMON_ALL_H
-#define LIBOPENCM3_PWR_COMMON_ALL_H
+#ifndef LIBOPENCM3_PWR_COMMON_V1_H
+#define LIBOPENCM3_PWR_COMMON_V1_H
 
 /**@{*/
 
@@ -126,7 +126,7 @@ END_DECLS
 #endif
 /** @cond */
 #else
-#warning "pwr_common_all.h should not be included explicitly, only via pwr.h"
+#warning "pwr_common_v1.h should not be included explicitly, only via pwr.h"
 #endif
 /** @endcond */
