+++ include/libopencm3/stm32/f1/pwr.h
@@ -31,7 +31,7 @@ LGPL License Terms @ref lgpl_license
 #ifndef LIBOPENCM3_PWR_H
 #define LIBOPENCM3_PWR_H
 
-#include <libopencm3/stm32/common/pwr_common_all.h>
+#include <libopencm3/stm32/common/pwr_common_v1.h>
 
 #endif
