+++ include/libopencm3/stm32/f3/pwr.h
@@ -41,7 +41,7 @@
 #ifndef LIBOPENCM3_PWR_H
 #define LIBOPENCM3_PWR_H
 
-#include <libopencm3/stm32/common/pwr_common_all.h>
+#include <libopencm3/stm32/common/pwr_common_v1.h>
 
 /* --- PWR_CR values ------------------------------------------------------- */
