+++ include/libopencm3/stm32/f4/pwr.h
@@ -34,7 +34,7 @@ LGPL License Terms @ref lgpl_license
 #ifndef LIBOPENCM3_PWR_H
 #define LIBOPENCM3_PWR_H
 
-#include <libopencm3/stm32/common/pwr_common_all.h>
+#include <libopencm3/stm32/common/pwr_common_v1.h>
 
 /*
  * This file extends the common STM32 version with definitions only
