+++ include/libopencm3/stm32/l1/pwr.h
@@ -36,7 +36,7 @@ LGPL License Terms @ref lgpl_license
 #ifndef LIBOPENCM3_PWR_H
 #define LIBOPENCM3_PWR_H
 
-#include <libopencm3/stm32/common/pwr_common_l01.h>
+#include <libopencm3/stm32/common/pwr_common_v2.h>
 
 /*
  * This file extends the common STM32 version with definitions only
