+++ lib/stm32/common/pwr_common_v2.c
