+++ lib/stm32/l0/Makefile
@@ -36,7 +36,7 @@ TGT_CFLAGS      += $(DEBUG_FLAGS)
 ARFLAGS		= rcs
 
 OBJS		= gpio.o rcc.o desig.o
-OBJS		+= pwr_common_all.o pwr_common_l01.o
+OBJS		+= pwr_common_v1.o pwr_common_v2.o
 OBJS		+= timer_common_all.o
 
 OBJS            += gpio_common_all.o gpio_common_f0234.o rcc_common_all.o
